Fetch error log on early failure
In case the pycalrisian pod responsible for executing a process failed at a very early stage, basically at start, we go and fetch the log file for this pod. In other cases, when issue happen later on, it usually fails to fetch such logs because of 403 due to ongoing deletion.